requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
async-timeout==5.0.1
blinker==1.4
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
idna==3.8
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.0
oauthlib==3.2.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==8.3.3
pytest-mock==3.14.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==5.4.1
rdkit==2024.3.4
redis==5.2.0
rich==13.8.1
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
systemd-python==234
tomli==2.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uvicorn==0.30.6
wadllib==1.3.6
watchfiles==0.24.0
websockets==13.0.1
zipp==1.0.0
